scanners: reorder again and add work-around for --as-needed brokeness
We want to have the objects first, then the libs and the target last. For
historic reasons ldflags and libs are merged :/. Ideally we'd like to turn
off as-needed in the docs (won't help here anyway), but I am not sure
wheter its a good idea to add "-Wl,--no-as-needed" to the linker command
(in regard to platform compat).