Ignore more decorators on functions
gtk-doc was explicitly ignoring only G_GNUC_*, which is good enough
for G_GNUC_MALLOC, G_GNUC_PRINTF etc. But for deprecation annotations,
we want to be able to change these on a per-module basis, so we
need to be able to wrap G_GNUC_DEPRECATED in, say, a GTK_DEPRECATED
macro. And it would be good if gtk-doc would not stumble over it.

https://bugzilla.gnome.org/show_bug.cgi?id=624001