gtkdoc-scan: recognise attribute-based deprecations
If we see something that looks like a macro for an attribute-based
deprecation (ie: some string matching /_DEPRECATED/) then mark the next
single declaration as <DEPRECATED/> in the -decl.txt. This avoids the
warning for missing deprecation guards.
Add Deprecated: tags to the doc-blobs of the existing tests to verify
that we don't get the warnings.

Fixes: #624001