make: create subdirectories of the content_files to the builddir
Fixup for "make: copy the files with their relative path included" also
by me.
That is it is fine to keep the subdir from the source to the build
directory, only it could not already have been generated. Do so.
Required by a followup patch for PackageKit which has SETUP_FILES
in a the "spec" directory. As nothing is generated there, it needs
action to create it in the build tree.

https://bugzilla.gnome.org/show_bug.cgi?id=720061