configure: search for xml catalog in XDG_DATA_DIRS
Detect the xml catalog file in either /etc/xml/catalog (currently the
only place we check) and also in XDG_DATA_DIRS/xml/catalog.

This will help find the XML catalog on systems that have it in
/usr/local/share/xml/catalog (like the BSDs) and will also help to find
it in the case that we've installed docbook ourselves as part of
bootstrapping jhbuild (like on Mac OS).

--with-xml-catalog is still supported and, if given, we will only
attempt to use the given path and not guess others.

https://bugzilla.gnome.org/show_bug.cgi?id=721228