gtkdoc-scan: fix regex for get_type() functions
The regexp /(void|)/ was apparently trying to look for "(void)" or "()",
but in fact, the parens here were only acting as (redundant) grouping,
so the expression would match any string containing either "void" or "",
which is every string.

It turns out that we don't want to look for parens anyway, since they
are already stripped out above.  We do want to look for exactly "void"
or "", however, so add ^ and $.

Unfortunately, some code above also replaces all whitespace surrounding
newlines with a single space character, and a trailing newline is left
at the end of the declaration by another regular expression above,
resulting in seeing "void " at this point.  Fix that expression up to
also clean up the newline at the end, by adding /s as a regexp flag.

This fixes the inappropriate matching of non-void functions, such as
g_io_extension_get_type().

https://bugzilla.gnome.org/show_bug.cgi?id=746118