Support highlight options
This commit allows gtk-doc users to supply a --src-lang option to
gtkdoc-fixxref so that the syntax highlight will be performed in the
correct language. The option is global, ie. applies to the whole project
built, as opposed to individual files. The default option is "c", which
is what used to be hardcoded in the gtkdoc-fixxref, so backwards
compatibility is maintained.

https://bugzilla.gnome.org/show_bug.cgi?id=749816