mkhtml/pdf: Fix quoting of search path
We need to do quoting using POSIX sh syntax, so this fix is uglier than simply
using bash arrays.

This still leaves the quoting of imgdirs in mkpdf and I can't see a non-terrible
way to quote that in POSIX sh syntax, but that is usually a relative path and is
less likely to have spaces.

https://bugzilla.gnome.org/show_bug.cgi?id=769125