mkhtml/pdf: fix out-of-srcdir builds
Fix how we assemble the --path param to xsltproc. We can only use --path once.
If there are multipel path we can separate them with a colon.

Also pass the search-path in dblatex through to xsltproc.

Finally from reading the xslt-sources we found out that the searchpath works
by appending only the filename part to each searchpath entry. Hence define
a path for the gobject example so that the source-file includes work.