Replace `match.groups(1)` with `match.group(1)`
halfline ran into a crash in gtk-doc when trying to cut an
accountsservice release, a TypeError pointing to this re.sub
call. Looking at it, the use of `match.groups(1)` is clearly
wrong. `match.groups()` returns a tuple consisting of *all*
the match subgroups; the argument it takes is the value to use
for subgroups which didn't capture anything (the default being
None). What the code here clearly actually *wants* is not that
tuple, but the contents of the first match subgroup only, as a
string. To get that you do `match.group(1)`. So, let's fix that.

There are two other occurrences of the same error later in the
file, so let's fix that too. If I'm reading it correctly, those
ones wouldn't have caused crashes, they would only cause the
block they're in not to work properly and produce "Can't
determine package for '(something)'" log messages even when it
should have worked (because 'package' will be the tuple, not the
subgroup, and will never be 'in' `OnlineMap` or `LocalMap`).

Note, these have been lying around for a long time, but the one
that causes the crash was not hit until 1.28, because of the
regex error fixed by b77d97b. Until that regex was fixed,
ReadDevhelp never worked on this codebase, so we never hit the
bug in ReadIndex. The crash might have happened with some other
codebase for which the ReadDevhelp regex *did* work, though.

Signed-off-by: Adam Williamson <awilliam@redhat.com>

https://bugzilla.gnome.org/show_bug.cgi?id=796012