gtkdoc-check: Don’t raise an exception if Makefile doesn’t exist
If running gtkdoc-check on a project built by Meson, neither Makefile.am
or Makefile will exist. We can set the gtkdoc-check configuration via
environment variables, which is fine; except that the get_variable()
code raises an exception if the variable is not found in the Makefile.
If the Makefile is missing, lines == [], which would cause the exception
to be raised.

Fix this by only grepping for the variable in the Makefile if it’s not
set in the environment.

Signed-off-by: Philip Withnall <withnall@endlessm.com>