scanner: avoid passing regex in command line
--ignore-decorator takes a regex such as `DEPRECATED_FOR\(.*\)` but that
leads to escaping nightmares and doesn't work when using Meson. The
ninja backend translates every `\` to `/` otherwise it breaks paths on
Windows, it is a know bug in Meson's ninja backend that cannot (easily)
be fixed.

Workaround this by allowing `DEPRECATED_FOR()` syntax and internally
translate that to the proper regex. That's the only use-case I'm aware
of.