Expand markdown in content files
Here is a quick proof-of-concept patch to recognize
.md files in expand_content_files, and run ConvertMarkDown
on them. It is not quite as useful as I had hoped, since
gtk-doc's markdown support is so limited. And it insists
on emitting <refsect2> elements. But its a start.

I tested this by putting test.md in expand_content_files,
and then have this in my main xml:
    <refentry>
      <refmeta><refentrytitle>Test</refentrytitle></refmeta>
      <refnamediv><refpurpose>markdown</refpurpose></refnamediv>
      <xi:include href="xml/test.md" />
    </refentry>