scangobj: fix -Wdouble-promotion warnings
If a project passes the -Wdouble-promotion CFLAG when compiling the
generated *scan.c file, then it emits some warnings. For example in the
Devhelp project:

```
Building documentation for devhelp-3
devhelp-3-scan.c: In function ‘describe_double_constant’:
devhelp-3-scan.c:595:39: warning: implicit conversion from ‘float’ to
‘gdouble’ {aka ‘double’} to match other operand of binary expression
[-Wdouble-promotion]
  595 | #define GTKDOC_COMPARE_FLOAT(x, y) (x <= y && x >= y
      |                                       ^~
devhelp-3-scan.c:608:12: note: in expansion of macro ‘GTKDOC_COMPARE_FLOAT’
  608 |   else if (GTKDOC_COMPARE_FLOAT (value, G_MAXFLOAT))
      |            ^~~~~~~~~~~~~~~~~~~~

[...]
```

An example where the -Wdouble-promotion CFLAG is used is with the
AX_COMPILER_FLAGS Autotools macro, part of autoconf-archive.

-----

In the diff, the `value` variable is of type gdouble.