scangobj: Support Visual Studio compilers
We make use of distutils to check what type the default compiler is, and we pass in CC=cl
if we detected Visual Studio to be the default compiler and CC is not set otherwise.  We
also set LD=link when we are using Visual Studio and when LD is not set.

Also set the compiler and linker flags appropriately when we are using Visual Studio when
running gtkdoc-scangobj.

Make sure we use forward slahes when retrieving the base_filename in scangobj.py so that
when we build the scanner program the program does not get confused on the paths of the
various files that is built using base_filename.