CI: Check that Meson tests can pass from an Autotools dist tarball
This allows the Autotools dist tarball to continue to be used as the
official release artifact, while still allowing distributors to try
Meson builds.

To help to debug this if it fails, the meson-build step records a
list of all files in git (ideally we'd use 'meson dist' but the
container image doesn't currently include git, so this is the next
best thing), while the meson-from-tarball step records a list of all
files in the tarball. They are in a normalized order that can be diff'ed.

Tests: #109
Signed-off-by: Simon McVittie <smcv@debian.org>