gtkdoc_uninstalled.py.in: Avoid unicode error on Windows
On Windows, if the gtk-doc sources were checked out or extracted or built under
the user's home directory, in c:\Users\<username>, the script will fail due to
a SyntaxError (unicode error) as Python interprets '\U' to be a prefix to an
unicode literal.

Fix this by prepending the strings that indicate the sourcedir and builddir
with an 'r', meaning that a raw string is being used.

Please note that the installed scripts are not affected as they use forward
slashes for the paths.