doap: update/improve the infos
The GTK stack is moving off gtk-doc, to use gi-docgen instead. So say
"GLib-based libraries".

https://gitlab.gnome.org/Infrastructure/gtk-web no longer hosts the
gtk-doc webpage, so point to GitLab instead for the homepage.

Mailing lists are discouraged now in GNOME, Discourse is preferred. But
there is no gtk-doc tag. So remove the mailing-list tag for now.

The programming language listed in the doap is for the primary language
the module is implemented in, which is currently Python.