ChangeLog: replace content by a message to look at the NEWS file
Worst case, the previous content of the ChangeLog is still kept in the
git history, of course.

Since there is still the Autotools build system, maybe a ChangeLog file
is still needed (if the gnu strictness is used). It can be looked at it
later, ideally the ChangeLog file should be removed completely.