gtkdocize: don't use docdir from environment
There is a logic bug in the initial configuration of gtkdocize, where it
replicates the docdir default assignment with the same idiom as the
srcdir default assignment, but docdir hasn't been assigned yet.

If docdir exists in the environment when gtkdocize is called then this
value is used as the source documentation directory, which is
unfortunate when $docdir is set alongside $prefix, so is /usr/share/doc.

As the intention at this stage of the script is to just set docdir to
srcdir, there's no need to look at the current value.

https://gitlab.gnome.org/GNOME/gtk-doc/-/merge_requests/73