Meson: a fix for test-gobject-mkhtml
tests/gobject/docs/Makefile.am has this line:

```
MKHTML_OPTIONS=--path=$(abs_top_srcdir)/tests/gobject/examples
```

It was missing in the Meson counterpart.

It fixes this problem:

```
warning: failed to load external entity "../../examples/gobject.c"
../xml/object.xml:139: element include: XInclude error : could not load ../../examples/gobject.c, and no fallback was found
```

The unit test in Meson was successful in the past, but we have not fully
found the root cause for the new failure, although this is related to a
change in libxml2. (Ideally a bisect should be done against libxml2, to
find the first commit that makes the gtk-doc test to fail).

Fixes https://gitlab.gnome.org/GNOME/gtk-doc/-/issues/150